Move test task and runtime to its own file.
Addresses #318
